+++ css_test.go
@@ -13,10 +13,12 @@ func Test(t *testing.T) {
 func Example() {
 	c := NewCSS()
 	c.Media = "@media screen"
+	c.Import("https://fonts.googleapis.com/css?family=Open+Sans")
 	c.Style("#x", "margin: 0 0")
 	c.WriteTo(os.Stdout)
 	// output:
 	// @media screen{
+	// @import url('https://fonts.googleapis.com/css?family=Open+Sans');
 	// #x {
 	// margin: 0 0;
 	// }